move logger to shared folder
