+++ src/main/deltachat/messagelist.js
@@ -1,5 +1,5 @@
 const C = require('deltachat-node/constants')
-const log = require('../../logger').getLogger('main/deltachat/messagelist')
+const log = require('../../shared/logger').getLogger('main/deltachat/messagelist')
 const { integerToHexColor } = require('./util')
 const filesizeConverter = require('filesize')
 const mime = require('mime-types')
