+++ src/main/logins.js
@@ -1,7 +1,7 @@
 const { join, basename } = require('path')
 const fs = require('fs-extra')
 const DeltaChat = require('deltachat-node')
-const logger = require('../logger')
+const logger = require('../shared/logger')
 const log = logger.getLogger('main/find_logins', true)
 const { escapeEmailForAccountFolder } = require('./deltachat/util')
