+++ src/main/state.js
@@ -1,6 +1,6 @@
 const appConfig = require('./application-config')
 const { EventEmitter } = require('events')
-const log = require('../logger').getLogger('main/state')
+const log = require('../shared/logger').getLogger('main/state')
 const { promisify } = require('util')
 
 const SAVE_DEBOUNCE_INTERVAL = 1000
