+++ src/renderer/components/message/messageFunctions.js
@@ -2,7 +2,7 @@ import path from 'path'
 import { remote, ipcRenderer, shell } from 'electron'
 // import C from 'deltachat-node/constants'
 
-import { getLogger } from '../../../logger'
+import { getLogger } from '../../../shared/logger'
 const log = getLogger('render/msgFunctions')
 /**
  * @typedef {{chatId: number,
